+++ packages/fedify/src/federation/middleware.ts
@@ -2349,7 +2349,7 @@ export class ContextImpl<TContextData> implements Context<TContextData> {
     if (!this.federation.manuallyStartQueue) {
       this.federation._startQueueInternal(this.data);
     }
-    this.federation.fanoutQueue.enqueue(
+    await this.federation.fanoutQueue.enqueue(
       message,
       { orderingKey: options.orderingKey },
     );